Move schema to a separate go module
This creates a schema submodule to remove dependencies from the
cdi package that is more commonly used.
