+++ schema/validate.go
@@ -14,10 +14,9 @@
    limitations under the License.
 */
 
-package validate
+package schema
 
 import (
-	"tags.cncf.io/container-device-interface/schema"
 	cdi "tags.cncf.io/container-device-interface/specs-go"
 )
 
@@ -27,7 +26,7 @@ const (
 )
 
 // WithSchema returns a CDI Spec validator that uses the given Schema.
-func WithSchema(s *schema.Schema) func(*cdi.Spec) error {
+func WithSchema(s *Schema) func(*cdi.Spec) error {
 	if s == nil {
 		return func(*cdi.Spec) error {
 			return nil
@@ -42,7 +41,7 @@ func WithSchema(s *schema.Schema) func(*cdi.Spec) error {
 // validator for it. If loading the schema fails a dummy validator is
 // returned.
 func WithNamedSchema(name string) func(*cdi.Spec) error {
-	s, _ := schema.Load(name)
+	s, _ := Load(name)
 	return WithSchema(s)
 }
 
@@ -50,9 +49,9 @@ func WithNamedSchema(name string) func(*cdi.Spec) error {
 // external JSON schema, or the default builtin one if the external one
 // fails to load.
 func WithDefaultSchema() func(*cdi.Spec) error {
-	s, err := schema.Load(DefaultExternalSchema)
+	s, err := Load(DefaultExternalSchema)
 	if err == nil {
 		return WithSchema(s)
 	}
-	return WithSchema(schema.BuiltinSchema())
+	return WithSchema(BuiltinSchema())
 }
